+++ romancal/lib/suffix.py
@@ -67,7 +67,7 @@ SUFFIXES_TO_ADD = [
 
 # Suffixes that are discovered but should not be considered.
 # Used by `find_suffixes` to remove undesired values it has found.
-SUFFIXES_TO_DISCARD = ["highlevelpipeline", "pipeline", "step"]
+SUFFIXES_TO_DISCARD = ["mosaicpipeline", "pipeline", "step"]
 
 
 # Calculated suffixes.