+++ CHANGES.rst
@@ -1,6 +1,12 @@
 0.15.2 (unreleased)
 ===================
 
+general
+-------
+
+- Rename highlevelpipeline to mosaicpipeline [#1249]
+
+
 source_catalog
 --------------
 - Add PSF photometry capability. [#1243]